requirements.txt
psycopg[binary]
psycopg2-binary
requests
flask
python-dotenv
